feat: implementado order repository

+++ src/main/java/br/selene/projectseleneback/domain/checkout/dto/ItemsCheckoutDTO.java
@@ -3,6 +3,6 @@ package br.selene.projectseleneback.domain.checkout.dto;
 public  record  ItemsCheckoutDTO(
         String name, 
         Integer quantity,
-        Integer unit_amount,
+        Long unit_amount,
         String imageURl
 ){}
